# Conflicts:
#	myopia-business/management/src/main/java/com/wupol/myopia/business/management/service/StatConclusionService.java

+++ myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/dto/stat/BasicStatParams.java
@@ -11,5 +11,5 @@ public class BasicStatParams {
     /** 占比 */
     private Float ratio;
     /** 数量 */
-    private Integer num;
+    private Long num;
 }
